+++ RuntimeLib/Runtime.h
@@ -219,7 +219,7 @@ RTBattleStyleClassFormulaDataRef RTRuntimeGetBattleStyleClassFormulaData(
 RTBattleStyleSlopeDataRef RTRuntimeGetBattleStyleSlopeData(
     RTRuntimeRef Runtime,
     Int32 SlopeID,
-    Int32 Penalty
+    Int64 Penalty
 );
 
 RTBattleStyleSlopeFormulaDataRef RTRuntimeGetBattleStyleSlopeFormulaData(
